+++ src/main/java/com/SWEProject/controllers/systemProductController.java
@@ -1,34 +1,24 @@
 package com.SWEProject.controllers;
 
-import com.SWEProject.Entities.Brand;
 import com.SWEProject.Entities.SystemProduct;
-import com.SWEProject.service.BrandService;
+
 import com.SWEProject.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/adminAPI")
+@RequestMapping("/systemProductAPI")
 @CrossOrigin(origins = "http://localhost:4200", allowedHeaders = "*")
 
-public class AdminController {
+public class systemProductController {
 
     @Autowired
-    ProductService ProductService;
-    @Autowired
-    BrandService BrandService;
-
+    ProductService productService;
 
     @PostMapping("/addSystemProduct")
     public boolean addSystemProduct(@RequestBody SystemProduct systemProduct)
     {
-        return ProductService.systemProductCheckAvailable(systemProduct);
-    }
-
-    @PostMapping("/addBrand")
-    public boolean addBrand(@RequestBody Brand brand)
-    {
-        return BrandService.brandCheckAvailable(brand);
+        return productService.systemProductCheckAvailable(systemProduct);
     }
 
 }
